alpine: fix build on current 'edge' version of Alpine
The python2 packet py-sphinx was removed from the edge repository, so we
use pip to install it localy for the package build process.

Also abuild assumes it is executed in a git folder and fails if no .git
is found in the folder or its parent folders, so we work around this, by
initializing an empty git repo with `git init`.

Signed-off-by: Juergen Werner <juergen@opensourcerouting.org>
Signed-off-by: harios <hari@niralnetworks.com>