Re-work newly added patches from frr-vyatta-6.0
Update the patches added in the parent commit so they apply against
FRR 7.2.

Much of the work required was in zebra-send-mpls-routes-to-FPM.patch.
Some of the re-factoring work of zebra_fpm.c done in the original
patch, to support processing of multiple queues, has been removed.
Since the original patch was written an additional queue was added
upstream (mac_q). Therefore we now use a similar pattern to integrate
the LSP queue.

VRVDR-48931